update nearest points

+++ 5thSemester/PAA/nearest_points/header_files/Resultado.hpp
@@ -1,7 +1,7 @@
 #pragma once
-
+#include <string>
 struct Resultado{
     std::string rotulo_primeiro;
     std::string rotulo_segundo;
-    int distancia;
+    float distancia;
 };
